requirements/prod.txt
asgiref==3.6.0
Django==3.2.16
django-allauth==0.54.0
django-ckeditor==6.5.1
django-environ==0.9.0
django-cleanup==7.0.0
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.4.0
pytz==2023.3
sqlparse==0.4.3
sorl-thumbnail==12.9.0
redis==4.5.4
celery==5.2.7
